# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "postit API",
    "description": "Self-hosted social publishing for a small team.",
    "contact": {
      "name": "asm",
      "email": "asm2do@outlook.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.7.0"
  },
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "The caller's own profile. Reachable while the account is still pending approval.",
        "description": "# Errors\n\nReturns 401 without a valid token, 403 for a disabled account.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token (`unauthenticated`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`account_disabled` (disabled or deleting account)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`; see `Retry-After`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "`unavailable` (signing keys not loaded yet)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidc": []
          }
        ]
      },
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Starts deleting the caller's own account.",
        "description": "# Errors\n\nReturns 403 while pending (or for a disabled or already-deleting account, which `Auth`\nrejects first), 422 when the display name does not match, 409 for the last active admin.",
        "operationId": "delete_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Deletion started"
          },
          "401": {
            "description": "No valid bearer token (`unauthenticated`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`account_disabled`, `account_pending`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`last_admin`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "`validation_failed`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`; see `Retry-After`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "`unavailable` (signing keys not loaded yet)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidc": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteMeRequest": {
        "type": "object",
        "required": [
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string",
            "description": "Must equal the caller's current display name."
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable problem code (the `code` member of every problem+json body).",
        "enum": [
          "unauthenticated",
          "account_pending",
          "account_disabled",
          "forbidden",
          "not_found",
          "request_timeout",
          "user_deleting",
          "last_admin",
          "idempotency_in_progress",
          "version_conflict",
          "payload_too_large",
          "validation_failed",
          "idempotency_key_reused",
          "precondition_required",
          "rate_limited",
          "internal",
          "unavailable"
        ]
      },
      "MeDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserDto"
          },
          {
            "type": "object",
            "properties": {
              "account_url": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "The identity provider's account page (`auth.oidc.account_url`), when configured."
              }
            }
          }
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "The RFC 9457 `application/problem+json` body every error response carries.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `X-Request-Id` of the failed request, for support."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`; `code` is the discriminator."
          }
        }
      },
      "RoleDto": {
        "type": "string",
        "enum": [
          "admin",
          "member"
        ]
      },
      "StatusDto": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "disabled",
          "deleting"
        ]
      },
      "UserDto": {
        "type": "object",
        "description": "A postit user as admins and the user themselves see it. Profile fields come from the identity provider.",
        "required": [
          "id",
          "email_verified",
          "display_name",
          "role",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "approved_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "approved_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "email_verified": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "role": {
            "$ref": "#/components/schemas/RoleDto"
          },
          "status": {
            "$ref": "#/components/schemas/StatusDto"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "oidc": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://idp.invalid/authorize",
            "tokenUrl": "https://idp.invalid/token",
            "scopes": {
              "email": "Email address",
              "offline_access": "Refresh tokens",
              "openid": "OpenID Connect sign-in",
              "profile": "Name and username"
            }
          }
        }
      }
    }
  }
}